+++ CHANGELOG.md
@@ -4,10 +4,16 @@
 
 Features:
 
-- 1: Creating bills
+- 1: Create and update bills
 
 ## Release 1.1.0
 
 Features:
 
-- 2: Creating bills
+- 2: Create competences on bill
+
+## Release 1.2.0
+
+Features:
+
+- 3: Create learning goals on competence